Split base class of equality engine manager to its own file.
