+++ src/theory/ee_manager_distributed.h
@@ -21,47 +21,15 @@
 #include <map>
 #include <memory>
 
-#include "theory/ee_setup_info.h"
-#include "theory/theory.h"
+#include "theory/ee_manager.h"
 #include "theory/uf/equality_engine.h"
 
 namespace CVC4 {
 
 class TheoryEngine;
+class SharedTermsDatabase;
 
 namespace theory {
-
-/**
- * This is (theory-agnostic) information associated with the management of
- * an equality engine for a single theory. This information is maintained
- * by the manager class below.
- *
- * Currently, this simply is the equality engine itself, which is a unique_ptr
- * for memory management purposes.
- */
-struct EeTheoryInfo
-{
-  EeTheoryInfo() : d_usedEe(nullptr) {}
-  /** The equality engine that the theory uses (if it exists) */
-  eq::EqualityEngine* d_usedEe;
-  /** The equality engine allocated by this theory (if it exists) */
-  std::unique_ptr<eq::EqualityEngine> d_allocEe;
-};
-
-/** Virtual base class for equality engine managers */
-class EqEngineManager
-{
- public:
-  virtual ~EqEngineManager() {}
-  /**
-   * Get the equality engine theory information for theory with the given id.
-   */
-  const EeTheoryInfo* getEeTheoryInfo(TheoryId tid) const;
-
- protected:
-  /** Information related to the equality engine, per theory. */
-  std::map<TheoryId, EeTheoryInfo> d_einfo;
-};
 
 /**
  * The (distributed) equality engine manager. This encapsulates an architecture
@@ -91,11 +59,18 @@ class EqEngineManagerDistributed : public EqEngineManager
    * This method is context-independent, and is applied once during
    * the lifetime of TheoryEngine (during finishInit).
    */
-  void finishInit();
-  /** get the master equality engine */
-  eq::EqualityEngine* getMasterEqualityEngine();
+  void initializeTheories() override;
+  /** get the model equality engine context */
+  context::Context* getModelEqualityEngineContext();
+  /** get the model equality engine */
+  eq::EqualityEngine* getModelEqualityEngine();
+  /** get the core equality engine */
+  eq::EqualityEngine* getCoreEqualityEngine() override;
 
  private:
+  /** Allocate equality engine that is context-dependent on c with info esi */
+  eq::EqualityEngine* allocateEqualityEngine(EeSetupInfo& esi,
+                                             context::Context* c);
   /** notify class for master equality engine */
   class MasterNotifyClass : public theory::eq::EqualityEngineNotify
   {
@@ -130,15 +105,21 @@ class EqEngineManagerDistributed : public EqEngineManager
     /** Pointer to quantifiers engine */
     QuantifiersEngine* d_quantEngine;
   };
-  /** Allocate equality engine that is context-dependent on c with info esi */
-  eq::EqualityEngine* allocateEqualityEngine(EeSetupInfo& esi,
-                                             context::Context* c);
   /** Reference to the theory engine */
   TheoryEngine& d_te;
   /** The master equality engine notify class */
   std::unique_ptr<MasterNotifyClass> d_masterEENotify;
   /** The master equality engine. */
   std::unique_ptr<eq::EqualityEngine> d_masterEqualityEngine;
+  /**
+   * A dummy context for the model equality engine, so we can clear it
+   * independently of search context.
+   */
+  context::Context d_modelEeContext;
+  /**
+   * The equality engine of the model.
+   */
+  std::unique_ptr<eq::EqualityEngine> d_modelEqualityEngine;
 };
 
 }  // namespace theory
